spike: data-items service refactor Refactor data_items component to move all logic to service so that in the future it might also be exposed to other components
